Add a default root_ca for OTA updates.
This is the currently active LetsEncrypt root CA[0], which seems like a
reasonable default that many folks may be using on their webservers and
therefore may "just work".

In the case that your server is secured with another CA, simply replace this file.

Note: this CA cert will change over time and will need updating before it
expires to prevent OTAs breaking (TODO!), but this is no worse than the
previous hard-coding in ota_cert.h really.

[0] sourced from https://letsencrypt.org/certificates/, specifically:
https://letsencrypt.org/certs/isrgrootx1.pem