Support board models and versioned firmwares
To support easily flashing different board configs (aka models) without having
to edit the source to change defines every time, set up platformio environments
for the different board configs (leds and neopixel) so you can simply use -e
MODEL to do the right thing with no editing required.

To support understanding what has been flashed, include a version number, build
timestamp and git revision hash in the built image, and display that in both
serial debug and MQTT messages. This works via a platformio extra_scripts pre
hook and associated python script. This hook also sets the generated firmware
filenames to include the model and version for convenience/clarity.

The versioning strategy is semver (for compatibility with esp32FOTA logic) with
the release version set in platformio.ini and used when the :release
environments are built. The default environments for each model, increment the
patch level of the release version and add a -BUILD_TIMESTAMP pre-release
suffix to the version they are built with, such that they can be forceOTA'd to
a device and a regular ota check will not cause them to be downgraded, but a
later release will cause them to be upgraded.