Use defaults for Wire initialization
Something in the newer platform code doesn't like the arguments we're passing
here, and fails to initialize the bus at all, whereas it works perfectly
without the arguments. The only argument that is potentially effectual here is
the clock, but that's already reset in the SCD30 specific code anyway, which
appears to be the main motivation for running at half the default speed, so
should be fine?