Support forced (aka specified file) OTA updates.
Add a new MQTT message and associated handlers for triggering an OTA with a
specified firmware file, rather than relying on the "is there a new version"
logic from parsing firmware.json.

This is useful for debug/testing over OTA when you don't want to be frequently
updating firmware.json.