add some OTA push helper utilities
These helper scripts ease the common actions involved in developing via OTA and
maintaining an OTA repository for the different models of board/firmware that
need to be flashed.

push-ota will build the specified environment and if it is a release, copy it
to the webserver and call the gen-firmwarejson.py script on that remote host to
update firmware.json accordingly. Optionally, if you also supply a device ID it
will then use trigger-ota to tell the device to check for a new version.

When operating on a non-release environment, the new firmware will be copied to
the webserver, but firmware.json is not updated. If you suply a device ID,
trigger-ota will then send a forced OTA request for that specific newly built
firmware to the device.