Store entire MqttMessage not just 4 bytes...
xQueueCreate takes the size of the item to store on the queue, which is
used by xQueueSendToBack to copy that number of bytes from the provided
pointer at queing time. So this should be sizeof(MqttMessage) not the
size of a pointer to MqttMessage (we can't store a pointer, because the
queued object is on the stack, not the heap)...

Despite the bug, this currently works because sizeof(MqttMessage) ==
sizeof(MqttMessgae *) atm - both are 4 bytes. But that won't be true
if/when more fields are added... (in the next commit).