Allow publishing of arbitrary status docs
With a helper to enable the simple usecase of just publishing a status
string (OTA use-case), but also allowing other use-cases (e.g.
reconnection) to send structured data rather than embedding # of
reconnects in a string.

This also avoids a race condition/potential loss of data if sendStatus
is called consecutively before the queue is processed.