Use heap not stack for the bulk of keygen memory
Generating a key requires a reasonable number of multi-kb buffers. If we
naively put those on the stack we have to enlarge the stack size for the
MQTT task, which permanently denies other tasks use of that memory, even
though it's going to be almost always unused in the MQTT stack given key
generation is a very rare operation. Allocating these buffers explicitly
on the heap means the memory is free for everything else to use the rest
of the time.