Support turning off lights below a threshold (#17)
Some users want to only display lights when action is required to
minimise distraction and visual pollution when all is well (e.g. in a
bedroom at night, or an office).

Address this request by adding a lower threshold (green), below which
the lights are turned off rather than green being shown anytime the
level is below the yellow threshold.

Also contains a minor alignment to the CO2 levels to be <= the
thresholds, rather than < to align with the iaq logic.

Arduino JSON size estimator reports the two new values do not increase
size above the current 1280 buffer value.