Support building/tagging variants on branches (#18)
To allow tagging the same version on different branches, we allow
version tags to be suffixed with '-name' (for differentiation in git),
and then strip the suffix following '-' in the versioning logic. This
allows multiple branches to generate the 'same' version in different
variants.

When detecting which branch we're building from we need to look for the
origin branch containing the tag, rather than the current method as git
builds in a detached HEAD state, and to make this possible we need to
set fetch-depth=0 on the checkout.

Finally, to tie it all together we include the branch name in the
generated version if the tag was not on the main branch.